requirements.txt
fastapi>=0.111.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
pyodbc>=5.0.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
prophet>=1.1.5
lightgbm>=4.0.0
xgboost>=1.7.6
joblib>=1.3.0
uvicorn[standard]>=0.30.0
